Refactor and add test to Icon component
